requirements.txt
bcrypt==3.1.7
bleach==2.1.3
certifi==2018.4.16
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-SQLAlchemy==2.3.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
packaging==17.1
passlib==1.7.2
psycopg2-binary==2.7.4
pyasn1==0.4.2
pyasn1-modules==0.2.1
python-decouple==3.3
requests==2.18.4
rsa==3.4.2
six==1.11.0
SQLAlchemy==1.2.7
webencodings==0.5.1
Werkzeug==0.14.1
